@@ -1,8 +1,8 @@
 import React from 'react'
 
-import Button from '@material-ui/core/Button'
-import Typography from '@material-ui/core/Typography'
-import Box from '@material-ui/core/Box'
+import Button from '@mui/material/Button'
+import Typography from '@mui/material/Typography'
+import Box from '@mui/material/Box'
 
 import { useUserValue } from '../../utils/contexts/UserContext'
 import wave1 from '../../assets/images/main-wave1.svg'
