@@ -1,7 +1,9 @@
 import React, { useMemo, useState } from 'react'
 import PropTypes from 'prop-types'
-import { Box, Button, makeStyles } from '@material-ui/core'
-import Grid from '@material-ui/core/Grid'
+import { Box, Button } from '@mui/material'
+// import { makeStyles } from '@mui/material/styles'
+import { makeStyles } from '@mui/styles'
+import Grid from '@mui/material/Grid'
 import moment from 'moment'
 import noImage from '../../../../assets/images/no-image.svg'
 import informationImage from '../../../../assets/images/fixedTagInformation.svg'
