@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react'
-import Typography from '@material-ui/core/Typography'
-import TextField from '@material-ui/core/TextField'
-import Box from '@material-ui/core/Box'
-import StreetviewIcon from '@material-ui/icons/Streetview'
+import Typography from '@mui/material/Typography'
+import TextField from '@mui/material/TextField'
+import Box from '@mui/material/Box'
+import StreetviewIcon from '@mui/icons-material/Streetview'
 import axios from 'axios'
 import Flag from '../../../Flag'
 import inputImg from '../../../../../../assets/images/input-icon.svg'
