@@ -1,17 +1,18 @@
 import React, { useEffect, useState } from 'react'
 import PropTypes from 'prop-types'
 import { usePlacesWidget } from 'react-google-autocomplete'
-import { makeStyles } from '@material-ui/core/styles'
-import Paper from '@material-ui/core/Paper'
-// import InputBase from '@material-ui/core/InputBase'
-import Divider from '@material-ui/core/Divider'
-import IconButton from '@material-ui/core/IconButton'
-import MenuIcon from '@material-ui/icons/Menu'
-import SearchIcon from '@material-ui/icons/Search'
-import CloseIcon from '@material-ui/icons/Close'
-import KeyboardReturnIcon from '@material-ui/icons/KeyboardReturn'
+// import { makeStyles } from '@mui/material/styles'
+import { makeStyles } from '@mui/styles'
+import Paper from '@mui/material/Paper'
+// import InputBase from '@mui/material/InputBase'
+import Divider from '@mui/material/Divider'
+import IconButton from '@mui/material/IconButton'
+import MenuIcon from '@mui/icons-material/Menu'
+import SearchIcon from '@mui/icons-material/Search'
+import CloseIcon from '@mui/icons-material/Close'
+import KeyboardReturnIcon from '@mui/icons-material/KeyboardReturn'
 import { useSnackbar } from 'notistack'
-// import MuiAlert from '@material-ui/lab/Alert'
+// import MuiAlert from '@mui/lab/Alert'
 import useMenu from '../../../../utils/hooks/useMenu'
 import SearchBarMenu from './SearchBarMenu'
 import Fillter from '../Filter/FilterFab'
