@@ -1,18 +1,19 @@
 import React, { useEffect } from 'react'
 import PropTypes from 'prop-types'
-import Typography from '@material-ui/core/Typography'
-import IconButton from '@material-ui/core/IconButton'
-import DialogContent from '@material-ui/core/DialogContent'
-import Grid from '@material-ui/core/Grid'
-import Dialog from '@material-ui/core/Dialog'
-import { makeStyles } from '@material-ui/core/styles'
-import Avatar from '@material-ui/core/Avatar'
-import Box from '@material-ui/core/Box'
-import CircularProgress from '@material-ui/core/CircularProgress'
-import { AppBar, Toolbar } from '@material-ui/core'
-import ArrowBackIcon from '@material-ui/icons/ArrowBack'
-import AssessmentIcon from '@material-ui/icons/Assessment'
-import EmailIcon from '@material-ui/icons/Email'
+import Typography from '@mui/material/Typography'
+import IconButton from '@mui/material/IconButton'
+import DialogContent from '@mui/material/DialogContent'
+import Grid from '@mui/material/Grid'
+import Dialog from '@mui/material/Dialog'
+// import { makeStyles } from '@mui/material/styles'
+import { makeStyles } from '@mui/styles'
+import Avatar from '@mui/material/Avatar'
+import Box from '@mui/material/Box'
+import CircularProgress from '@mui/material/CircularProgress'
+import { AppBar, Toolbar } from '@mui/material'
+import ArrowBackIcon from '@mui/icons-material/ArrowBack'
+import AssessmentIcon from '@mui/icons-material/Assessment'
+import EmailIcon from '@mui/icons-material/Email'
 import useUserDetail from '../../../../utils/hooks/useUserDetail'
 import { useUserValue } from '../../../../utils/contexts/UserContext'
 
