@@ -1,5 +1,4 @@
-// import { createMuiTheme } from '@material-ui/core/styles'
-import { createTheme } from '@material-ui/core/styles'
+import { createTheme } from '@mui/material/styles'
 
 const ThemeSettings = {
   palette: {
